Launch background after launcher.
This introduces a safer sequence for initialization of the HTML5
shell. The sequence will be launching html5-launcher and then
html5-background, with 1 second waits.

This is to ensure we do not send a request to launch two applications
in a very short timespan, that is not working because of WAM AGL
service not supporting that. So it is essentially a workaround for that
problem.

To improve launch experience, background is the last to be launched,
as it reduces the number of visible relayouts.

Bug-AGL: SPEC-3982
Change-Id: Icfdddb18edbda06e7ece1132f08a9fd469a469a8
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>